+++ database/migrations/2025_01_17_211530_modify_archivos_proyecto_add_pre.php
@@ -26,10 +26,12 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('archivos_proyecto', function (Blueprint $table) {
-            $table->dropForeign(['pre_proyecto_id']);
-            $table->dropColumn('pre_proyecto_id');
-        });
+        if (Schema::hasColumn('archivos_proyecto', 'pre_proyecto_id')) {
+            Schema::table('archivos_proyecto', function (Blueprint $table) {
+                $table->dropForeign(['pre_proyecto_id']);
+                $table->dropColumn('pre_proyecto_id');
+            });
+        }
 
     }
 };
