+++ database/migrations/2025_02_27_183236_add_flag_tallas_table.php
@@ -21,8 +21,10 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('categorias', function (Blueprint $table) {
-            $table->dropColumn('flag_tallas');
-        });
+        if (Schema::hasColumn('categorias', 'flag_tallas')) {
+            Schema::table('categorias', function (Blueprint $table) {
+                $table->dropColumn('flag_tallas');
+            });
+        }
     }
 };
