+++ database/migrations/2025_02_27_205901_add_name_id_contacto_direcciones_entrega_table.php
@@ -21,8 +21,10 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('direcciones_entrega', function (Blueprint $table) {
-            $table->dropColumn('nombre_empresa');
-        });
+        if (Schema::hasColumn('direcciones_entrega', 'nombre_empresa')) {
+            Schema::table('direcciones_entrega', function (Blueprint $table) {
+                $table->dropColumn('nombre_empresa');
+            });
+        }
     }
 };
