+++ database/migrations/2025_03_25_182018_add_url_and_last_uploaded_file_id_to_pedido_table.php
@@ -26,8 +26,10 @@ return new class extends Migration
      */
     public function down(): void
     {
+     if (Schema::hasColumn('pedido', 'url')) {
         Schema::table('pedido', function (Blueprint $table) {
             $table->dropColumn(['url', 'last_uploaded_file_id']);
         });
+      }
     }
 };
