+++ database/migrations/2025_03_31_170727_add_ind_activo_to_table_tallas.php
@@ -22,9 +22,12 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('tallas', function (Blueprint $table) {
-            //
-            $table->dropColumn('ind_activo');
-        });
+
+        if (Schema::hasColumn('tallas', 'ind_activo')) {
+            Schema::table('tallas', function (Blueprint $table) {
+                //
+                $table->dropColumn('ind_activo');
+            });
+        }
     }
 };
