+++ database/migrations/2025_04_04_191754_add_flag_aprobar_sin_fechas_to_pedido_table.php
@@ -24,8 +24,10 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('pedido', function (Blueprint $table) {
-            $table->dropColumn('flag_aprobar_sin_fechas');
-        });
+        if (Schema::hasColumn('pedido', 'flag_aprobar_sin_fechas')) {
+            Schema::table('pedido', function (Blueprint $table) {
+                $table->dropColumn('flag_aprobar_sin_fechas');
+            });
+        }
     }
 };
