+++ database/migrations/2025_04_07_180239_add_tipo_to_tareas_table.php
@@ -23,8 +23,10 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('tareas', function (Blueprint $table) {
-            $table->dropColumn('tipo');
-        });
+        if (Schema::hasColumn('tareas', 'tipo')) {
+            Schema::table('tareas', function (Blueprint $table) {
+                $table->dropColumn('tipo');
+            });
+        }
     }
 };
