+++ database/migrations/2025_04_21_113914_add_flag_armado_preproyecto.php
@@ -24,8 +24,10 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('pre_proyectos', function (Blueprint $table) {
-            $table->dropColumn('flag_armado');
-        });
+        if (Schema::hasColumn('pre_proyectos', 'flag_armado')) {
+            Schema::table('pre_proyectos', function (Blueprint $table) {
+                $table->dropColumn('flag_armado');
+            });
+        }
     }
 };
