+++ database/migrations/2025_05_05_155820_add_letra_to_layout_elementos_table.php
@@ -13,8 +13,10 @@ return new class extends Migration
     }
 
     public function down(): void {
+    if (Schema::hasColumn('layout_elementos', 'letra')) {
         Schema::table('layout_elementos', function (Blueprint $table) {
             $table->dropColumn('letra');
         });
+    }
     }
 };
