+++ database/migrations/2025_05_29_115832_add_user_can_sel_preproyectos_to_users_table.php
@@ -15,7 +15,9 @@ return new class extends Migration {
     public function down(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('user_can_sel_preproyectos');
+            if (Schema::hasColumn('users', 'user_can_sel_preproyectos')) {
+                $table->dropColumn('user_can_sel_preproyectos');
+            }
         });
     }
 };
