+++ database/migrations/2025_06_05_155410_add_subordinados_to_users_table.php
@@ -17,7 +17,9 @@ return new class extends Migration
     public function down()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('subordinados');
+            if (Schema::hasColumn('users', 'subordinados')) {
+                $table->dropColumn('subordinados');
+            }
         });
     }
 };
